requirements.txt
ccxt
pandas
plotly
python-binance